chore: Import os module in settings.py
This commit adds an import statement for the `os` module in the `settings.py` file. The `os` module is required to access environment variables. This change allows the code to use the `os.environ.get()` function to retrieve the value of the `KAFKA_GROUP_ID` environment variable, providing more flexibility in setting the group ID for Kafka.

Note: This commit message follows the established convention of using the "chore" type for non-functional changes and providing a clear and concise description of the changes made.