chore: Update business logic in Python.KafkaDemoBP
This commit updates the business logic in the `Python.KafkaDemoBP` class. The changes include deserializing the message to a `Product` object, parsing the message to a `Product` object, randomly choosing the message type, and sending the message to the `Python.LocalStorageBO` business operation.

Note: This commit message follows the established convention of using the "chore" type for non-functional changes and providing a clear and concise description of the changes made.